refactor(oxlint): split rule-maps.ts into external-plugin-rules + react-doctor-rules
`oxlint/rule-maps.ts` was mixing two unrelated concerns: 4 hand-maintained
maps of rules adopted from upstream plugins (react, jsx-a11y,
react-hooks-js, you-might-not-need-effect) and 5 react-doctor-derived
constants computed from `plugin.rules`. Splits them into purpose-named
sibling files and deletes the misnamed `rule-maps.ts`:

- **`oxlint/external-plugin-rules.ts`** — the 4 upstream-plugin maps.
  Pure data, no derivation. Loaded by `createOxlintConfig` to merge into
  the generated config.

- **`oxlint/react-doctor-rules.ts`** — the 5 per-framework maps PLUS
  `ALL_REACT_DOCTOR_RULE_KEYS` and `FRAMEWORK_SPECIFIC_RULE_KEYS`. All
  derived at module load from the colocated `defineRule({...})` metadata
  on each rule. The two `*_RULE_KEYS` sets now derive **directly** from
  `plugin.rules` instead of going through the per-framework maps, which
  cuts one indirection and makes the source of truth explicit.

Public API unchanged — `oxlint-config.ts` re-exports the same names from
their new locations, so consumers (`run-oxlint.ts`,
`merge-and-filter-diagnostics.ts`, `eslint-plugin.ts`) don't move.

Verified: byte-identical 262-line diagnostic snapshot from `runOxlint`
against `basic-react`, `nextjs-app`, and `tanstack-start-app` fixtures
vs `origin/main`. 734 / 734 tests pass.
